code/requirements.txt
azure-common==1.1.28
azure-core==1.23.1
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
greenlet==1.1.2
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
pycparser==2.21
pyodbc==4.0.32
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
urllib3==1.26.9
